Renamed sim_data_gen to generators.
Generators seemed a more appropriate name. We'll see if that doesn't
change again going forward
