--- data/sim_data_gen/src/sim_data_gen/__init__.py
@@ -1,2 +1,0 @@
-def hello() -> str:
-    return "Hello from sim-data-gen!"
